Updates CheckingAccount into an aggregate root
I will see about the users later.
Want to try it to be as close as YAGNI as possible.

Remain the deposit and withdraw method to implements with domain events.
I need to strat to work on the ledger for that too.
Deposits and withdrawals are operations, interface needed there.
On operation has an amount and a date of ? occuredOn ?
Then the ledger is a list of operations orderd by date.