Changes the namespaces
I've started by changing the Symfony "App" namespaces by "ElyAccount"
I thought it was nice to configure it to be lesss generic and gave it
the name of the application.

But Symfony use the "App" namespace here and there, nothing that I can't
changes but it's a bit of extra work.
Plus it's kind of nice to have a namspace just for the application, it's
separates the domain even more.